NO-JIRA: tests(odh-nbc): update to the latest route crd
There's an issue I encountered that when I was prepopulating some routes in a test,
when I was setting the `status`, I was getting errors because status was not declared as subresource in the CRD we were using before.

I am therefore updating the CRD we use in tests using one that comes directly from OpenShift.