Synces kustomization.yaml namespace with what we have in upstream
The odh/rhoai operator overrides the namespace values anyway, so we
should be okay. With this in sync with upstream, we reduce the code
difference and will able the e2e tests to work as are defined now.

We may want to rethink the e2e test env preparation in the future
eventually.