[Add] Propagate labels from NB to StatefulSets
This commit propagates labels from Notebooks to statefulsets.
This is required to enable NBs to be managed by Kueue. When a user
adds the label to point a NB to the right LocalQueue where it
needs to be admitted, it needs to propagate to the underlying
StatefulSet, so that Kueue can manage it.

Signed-off-by: Varsha Prasad Narsing <varshaprasad96@gmail.com>
(cherry picked from commit f6abf13cf8032fe273fedf8a413f89805b5f4fc8)