NO-JIRA: chore(dev): log error when imagestream is found but could not be fetched
This is something that should not normally happen, but it happened to me when I was working with faked imagestreams in tests and the lack of logging confused me.

I had to resort to running the webhook in debugger to understand what was going on.

With this log message, such error should be easier to debug, without stepping through the code.