NO-JIRA: chore(gh): permit filling blank github issues on the opendatahub-io/kubeflow repo again
Kubeflow upstream disabled filling new issues on kubeflow/kubeflow and we inherited the setting for our own repos.

Undo this. It's great to be able to jot down ideas in GH issue tracker, when the ticket is not ready for the ceremonies heavy agile process we have in our company Jira.