[RHOAIENG-33609] fix route search to work also for generated routes
In the `createOAuthClient` function, the search for the appropriate
route for the workbench didn't work when the generated route was created
before. We're now searching based on the appropriate labels as is used
for other resources in this case when generated name is used instead of
the regular one.

A new test was added and I had to increase the duration timeout as due
to more notebook instances being in the testing environment now, the
10seconds wasn't always enough for the reconciliation loop to finish in
time for the test.

This is a followup of the #539.