[RHOAIENG-34310] replace oauth-proxy with kube-rbac-proxy sidecar (#698)
* [RHOAIENG-34310] replace oauth-proxy with kube-rbac-proxy sidecar

Due to the unified approach to authentication, we are replacing the
oauth-proxy with the kube-rbac-proxy sidecar for the Notebook instance.
As such, we'll be having a SSO in RHOAI and Notebook instances will be
checking the actual authorization bits for each request.

This also requires replace of the original route.openshift.io/v1
resource with the gateway.networking.k8s.io/v1 resource to work smoothly
with the Gateway API approach now.

* configmap ownership defined for reconciliation
* add dspa new permissions