[RHOAIENG-38009] Create HTTPRoute in a central namespace
With our previous approach where we created HTTPRoute in a users
namespace and accept this route in our Gateway instance, this setup
could lead to eventual unprivileged user access to the workbench
instance. We updated our approach so that odh-notebook-controller is now
creating the HTTPRoute instances in a central namespace (and only from
this namespaces these are accepted by Gateway instance) adding also a
ReferenceGrant instance to permit the access to the Services in the
users namespace.

(cherry picked from commit 22af0485c2cecad30b96f746793abcc6d35d097f)