[RHOAIENG-24545] Assure proper pipelines runtime config/secret is present
Before this patch, the Secret that contains the pipelines runtime
configuration for the Jupyter IDE wasn't created in time for the very
first workbench started in a newly created namespace. Workaround was
either to restart the workbench or re-create it. Now this should work
just fine even for the first workbench start.